requirements.txt
flask==2.3.3
pandas==2.1.0
numpy==1.26.0
scikit-learn==1.3.0
plotly==5.15.0
joblib==1.3.2
imbalanced-learn==0.11.0
openpyxl==3.1.2
werkzeug==2.3.7
openai==1.3.0
mlflow==2.7.1
gunicorn==21.2.0
flask-login==0.6.2
flask-sqlalchemy==3.0.5
flask-wtf==1.1.1
pymysql==1.0.3
mysqlclient==2.1.1
email-validator==2.0.0
cryptography==40.0.0
sqlalchemy==1.4.49
alembic==1.6.0
wtforms==3.0.0
docker==6.1.3
requests>=2.25.0
torch==2.0.0
torchvision==0.15.1
requests-toolbelt==0.10.1
urllib3==1.26.16
azure-storage-blob==12.18.2
azure-identity==1.13.0
azure-keyvault-secrets==4.6.0
